+++ NEWS.md
@@ -1,4 +1,16 @@
 
+'replyr' 0.2.3 2017-02-20
+ 
+ * Move "let()" to wrapr https://github.com/WinVector/wrapr .
+ * add replyr_coalesce.
+ * Add many work-arounds for different remote data stores.
+ * Make minimal Spark version: 2.0.0.
+ * replyr_bind_rows reorder and intersect column names.
+ 
+'replyr' 0.2.2 2017-02-04
+ 
+ * Add Debug* functions.
+
 'replyr' 0.2.1 2017-01-21
 
  * Excise direct use of lazyeval.